mimesis-producer/requirements.txt
kafka-python
mimesis